ast: move parse module back into parent
